does it fix Karma circular dependencies?
