+++ editor/karma-setup.js
@@ -1,3 +1,10 @@
+/**
+ * importing the shared entry point imports file, which should in theory mean that
+ * the real running Editor and the test editors resolve dependencies in the same order,
+ * which should in theory prevent diverging circular dependencies
+ */
+import './src/templates/editor-entry-point-imports'
+
 // // The matchers API
 import expect from 'expect'
 import * as BrowserFS from 'browserfs'
